Restructure components packaging
